--- tests/examples/conftest_functions.py
@@ -1,21 +1,0 @@
-import pytest
-
-
-@pytest.fixture
-def string_value():
-    return "test_value"
-
-
-@pytest.fixture
-def number_value():
-    return 123
-
-
-@pytest.fixture
-def dict_value():
-    return {"key": "value", "number": 42}
-
-
-@pytest.fixture
-def base_url():
-    return "https://jsonplaceholder.typicode.com"
